+++ packages/cli-core/src/package-manager-controller/lock-file-reader/pnpm_lock_file_reader.ts
@@ -46,7 +46,7 @@ export class PnpmLockFileReader implements LockFileReader {
 
         dependencies.push({ name: dependencyName, version: dependencyVersion });
       }
-    } catch (error) {
+    } catch {
       printer.log(
         `Failed to get lock file contents because ${pnpmLockPath} does not exist or is not parse-able`,
         LogLevel.DEBUG