+++ scripts/components/dependencies_validator.test.ts
@@ -1,7 +1,7 @@
 import { before, describe, it, mock } from 'node:test';
 import assert from 'node:assert';
 import { DependenciesValidator } from './dependencies_validator.js';
-import { ExecaChildProcess } from 'execa';
+import { ExecaMethod } from 'execa';
 import fsp from 'fs/promises';
 import { fileURLToPath } from 'url';
 import { glob } from 'glob';
@@ -10,7 +10,7 @@ void describe('Dependency validator', () => {
   const execaMock = mock.fn(() => {
     return {
       stdout: '',
-    } as unknown as ExecaChildProcess;
+    } as unknown as ReturnType<ExecaMethod>;
   });
 
   before(async () => {
@@ -23,7 +23,7 @@ void describe('Dependency validator', () => {
     execaMock.mock.mockImplementation(() => {
       return {
         stdout: testNpmOutput,
-      } as unknown as ExecaChildProcess;
+      } as unknown as ReturnType<ExecaMethod>;
     });
   });
 
@@ -259,7 +259,7 @@ void describe('Dependency validator', () => {
     const packagePaths = await glob(
       'scripts/components/test-resources/inter-repo-dependency-version-consistency-test-packages/*'
     );
-    const validator = await new DependenciesValidator(
+    const validator = new DependenciesValidator(
       packagePaths,
       {},
       [],
@@ -292,7 +292,7 @@ void describe('Dependency validator', () => {
         await new DependenciesValidator(
           packagePaths,
           {},
-          [['aws-cdk', 'aws-cdk-lib']],
+          [['dep-1', 'dep-2']],
           [],
           execaMock as never
         ).validate();
@@ -301,8 +301,8 @@ void describe('Dependency validator', () => {
         assert.ok(
           err.message.includes('should be declared using same version')
         );
-        assert.ok(err.message.includes('aws-cdk'));
-        assert.ok(err.message.includes('aws-cdk-lib'));
+        assert.ok(err.message.includes('dep-1'));
+        assert.ok(err.message.includes('dep-2'));
         return true;
       }
     );