+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,77 @@
 # @aws-amplify/backend-cli
 
+## 1.4.12
+
+### Patch Changes
+
+- dafb530: Bumps [is-ci](https://github.com/watson/is-ci) from 3.0.1 to 4.1.0.
+- Updated dependencies [dafb530]
+- Updated dependencies [ca9d732]
+- Updated dependencies [af90a9b]
+- Updated dependencies [aec5e08]
+  - @aws-amplify/platform-core@1.6.4
+  - @aws-amplify/backend-deployer@1.1.19
+  - @aws-amplify/cli-core@1.4.0
+
+## 1.4.11
+
+### Patch Changes
+
+- cceadef: updated tests and constructor for profileController
+- Updated dependencies [b5f785f]
+- Updated dependencies [ce8e4fc]
+  - @aws-amplify/platform-core@1.6.3
+
+## 1.4.10
+
+### Patch Changes
+
+- de90deb: Added catch blocks for EACCES Errors
+- 6faa636: Add data and ai packages to ampx info output
+- bc07307: Update code with Eslint@8 compliant
+- Updated dependencies [de90deb]
+- Updated dependencies [bc07307]
+  - @aws-amplify/client-config@1.5.7
+  - @aws-amplify/backend-deployer@1.1.17
+  - @aws-amplify/form-generator@1.0.4
+  - @aws-amplify/platform-core@1.6.2
+  - @aws-amplify/cli-core@1.2.4
+
+## 1.4.9
+
+### Patch Changes
+
+- 99bb9a4: enforce branch option implying app id for generate commands
+- Updated dependencies [a7ae994]
+- Updated dependencies [ec7826d]
+  - @aws-amplify/backend-deployer@1.1.16
+  - @aws-amplify/client-config@1.5.6
+  - @aws-amplify/backend-secret@1.1.6
+  - @aws-amplify/platform-core@1.6.1
+
+## 1.4.8
+
+### Patch Changes
+
+- ff2f2ce: fixed violations to cause propagation lint rule
+- Updated dependencies [ff2f2ce]
+- Updated dependencies [ff2f2ce]
+  - @aws-amplify/backend-deployer@1.1.15
+  - @aws-amplify/schema-generator@1.2.7
+  - @aws-amplify/cli-core@1.2.3
+  - @aws-amplify/sandbox@1.2.10
+
+## 1.4.7
+
+### Patch Changes
+
+- 2dab201: Report cdk versions
+- Updated dependencies [2dab201]
+  - @aws-amplify/backend-deployer@1.1.14
+  - @aws-amplify/cli-core@1.2.2
+  - @aws-amplify/platform-core@1.6.0
+  - @aws-amplify/plugin-types@1.8.0
+
 ## 1.4.6
 
 ### Patch Changes