+++ packages/client-config/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @aws-amplify/client-config
 
+## 1.5.7
+
+### Patch Changes
+
+- de90deb: Added catch blocks for EACCES Errors
+- Updated dependencies [bc07307]
+  - @aws-amplify/platform-core@1.6.2
+
+## 1.5.6
+
+### Patch Changes
+
+- a7ae994: wrap more instances of ENOENT type errors
+- Updated dependencies [ec7826d]
+  - @aws-amplify/platform-core@1.6.1
+
 ## 1.5.5
 
 ### Patch Changes