+++ packages/integration-tests/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @aws-amplify/integration-tests
 
+## 0.6.3
+
+### Patch Changes
+
+- 3298c3a: update naming of storage access policies to also use stack
+
+## 0.6.2
+
+### Patch Changes
+
+- ff2f2ce: fixed violations to cause propagation lint rule
+
 ## 0.6.1
 
 ### Patch Changes