+++ packages/sandbox/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @aws-amplify/sandbox
 
+## 1.2.11
+
+### Patch Changes
+
+- b359bc6: move from importing the default export from the commonjs distribution of parcle/watcher to named export
+
+## 1.2.10
+
+### Patch Changes
+
+- ff2f2ce: updates bootstrap message
+- Updated dependencies [ff2f2ce]
+  - @aws-amplify/backend-deployer@1.1.15
+  - @aws-amplify/cli-core@1.2.3
+
 ## 1.2.9
 
 ### Patch Changes